% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_quadreg_2018.R
\name{check_n}
\alias{check_n}
\title{Checks if enough data points are avaliable}
\usage{
check_n(trend_variable, min_n = 7)
}
\arguments{
\item{trend_variable}{vector with value}

\item{min_n}{min value for trendanalysis}
}
\description{
Checks if enough data points are avaliable
}
