% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bp.R
\name{compute_bp}
\alias{compute_bp}
\title{Search for breakpoints using the segmented package. Picking the best model with BIC}
\usage{
compute_bp(
  dat,
  varname,
  type = "BIC",
  control = seg.control(n.boot = 50, it.max = 1000, alpha = 0.05),
  psi = NULL,
  npsi = 2,
  msg = TRUE,
  plot = TRUE
)
}
\arguments{
\item{dat}{Data frame to be analysed}

\item{type}{Currently set to BIC only}

\item{control}{Argument used in segmented. Defaults to
seg.control(n.boot = 50, it.max = 1000, alpha = 0.1)
The alpha-value is set quit high to avoid finding break points near the edges
See also ?seg.control for more parameters}

\item{psi}{Psi can be a numeric vector of guesses on breakpoints passed on to the segmented function
Segmented takes the psi values as starting points to find segments in the data. Defaults to NULL}

\item{npsi}{A named vector or list meaning the number (and not locations) of breakpoints to be estimated.
The starting values will be internally computed via the quantiles or equally spaced values,
as specified in argument quant in seg.control. npsi can be missing and npsi=1
is assumed for all variables specified in seg.Z. If psi is provided, npsi is ignored.}

\item{msg}{logical. Indicates if results are displayed}

\item{plot}{Plots a graph using plotly}
}
\description{
The package segmented is used to analyse data for break points.
Breakpoints are chosen by least BIC and are then, if present, are plotted in
a comprehensive graph using plotly
}
\details{
The package segmented provides a tool to determine segmented relationships in regression models
with breakpoints / changepoints (with possibly random effects) estimation
}
