% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_quadreg_2018.R
\name{compute_trend}
\alias{compute_trend}
\title{Calculates the trend for data and variables
Die Funktion checkt ob genug Daten vorliegen und laesst dann das Modell rechnen.}
\usage{
compute_trend(dat, varname, log_trans = FALSE, calc_infl_obs = TRUE)
}
\arguments{
\item{dat}{dataset}

\item{varname}{colnames of dataset}

\item{log_trans}{defaults to FALSE and indicates whether variables of data
are log transformed or not}

\item{calc_infl_obs}{logical}
}
\value{
Eine Liste mit Trendgrafike, Variablenname, Modell und Text zum Ergebnis
}
\description{
Calculates the trend for data and variables
Die Funktion checkt ob genug Daten vorliegen und laesst dann das Modell rechnen.
}
