% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CookD_gls.R
\name{CookD_gls}
\alias{CookD_gls}
\title{Modified version of predictmeans::CookD}
\usage{
CookD_gls(
  model,
  group = NULL,
  plot = TRUE,
  idn = 3,
  newwd = TRUE,
  used_formula,
  cor_structure,
  data
)
}
\arguments{
\item{model}{
Model object returned by \code{aov}, \code{lm}, \code{glm}, \code{gls}, \code{lme}, and \code{lmer}. 
}

\item{group}{
Name (in "quotes") for indicating how observations are deleted for Cook's distance calculation. If \code{group!=NULL} then deletions will be along levels of \code{group} variable, otherwise, will be along individual observations.
}

\item{plot}{
  A logical variable; if it is true, a plot of Cook's distance will be presented. The default is TRUE.
}

\item{idn}{
  An integer indicating the number of top Cook's distances to be labelled in the plot. The default value is 3.
}

\item{newwd}{
A logical variable to indicate whether to print graph in a new window. The default value is TRUE.
}

\item{used_formula}{Formula used when fitting the model}

\item{cor_structure}{correlation structure used when fitting the}

\item{data}{data used for the fit}
}
\value{
vector with Cook's distance
}
\description{
Modified version of predictmeans::CookD
}
\details{
Rather quick and dirty method; not exported function from package
predictmeans is used. Maybe better implementation if scoping is adapted.
Copy of predictmeans::CookD, but with ability to include the used formula
and correlation structure so that the update function works. The model needs
to be fitted with ML.
}
