% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phi.R
\name{extract_phi}
\alias{extract_phi}
\title{Extract phi from the gls model}
\usage{
extract_phi(mod)
}
\arguments{
\item{mod}{gls model}
}
\value{
phi parameter of correlation
}
\description{
Extract phi from the gls model
}
