% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trend.R
\name{extract_trend.gls}
\alias{extract_trend.gls}
\title{Extract trend from gls-objects}
\usage{
\method{extract_trend}{gls}(model, alpha = 0.05)
}
\arguments{
\item{model}{gls model}

\item{alpha}{level of significance}
}
\value{
String with trendresults from gls
}
\description{
Extract trend from gls-objects
}
