% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trend.R
\name{extract_trend.lm}
\alias{extract_trend.lm}
\title{Extract trend from lm-objects}
\usage{
\method{extract_trend}{lm}(model, alpha = 0.05)
}
\arguments{
\item{model}{lm model}

\item{alpha}{Level of significance}
}
\value{
String with trend results from lm
}
\description{
Extract trend from lm-objects
}
