% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trend.R
\name{extract_trend}
\alias{extract_trend}
\title{Method for extract trends from model}
\usage{
extract_trend(model, alpha, ...)
}
\arguments{
\item{model}{gls or lm model}

\item{alpha}{Level of significance}

\item{...}{zur Time nicht benutzt}
}
\value{
String with trend results
}
\description{
Method for extract trends from model
}
