% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_quadreg_2018.R
\name{extrahiere_phi}
\alias{extrahiere_phi}
\title{Extrahiere phi aus gls - Modell}
\usage{
extrahiere_phi(mod)
}
\arguments{
\item{mod}{gls-Modell}
}
\value{
phi Parameter der Korrelation
}
\description{
Extrahiere phi aus gls - Modell
}
