% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_quadreg_2018.R
\name{extrahiere_trend}
\alias{extrahiere_trend}
\title{Extrahiere Trend aus Modell}
\usage{
extrahiere_trend(model, alpha, ...)
}
\arguments{
\item{model}{gls oder lm Modell}

\item{alpha}{Signifikanzniveau}

\item{...}{zur Zeit nicht benutzt}
}
\value{
String mit Trendergebnis
}
\description{
Extrahiere Trend aus Modell
}
