% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Calculate the model}
\usage{
fit_trend(dat, varname, alpha = 0.05)
}
\arguments{
\item{dat}{data.frame with data (Colnames Time, ID must be available
the first column is taken to be a continuous time series)}

\item{varname}{String with variablename}

\item{alpha}{Alphavalue for area of rejection}
}
\value{
lm or gls - model (linear, ML)
}
\description{
Calculate the model
}
