% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{Determine trend information using Sen's methodology and Kendall tests}
\usage{
get_trend(y)
}
\arguments{
\item{y}{a vector of data for which the trend anlysis will be undertaken.}
}
\value{
a data frame.
}
\description{
Using a vector of data, The Thiel-Sen method for trend determination will be carried out with a Kendall test for significance.
}

