% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_quadreg_2018.R
\name{infl_obs}
\alias{infl_obs}
\title{Bestimme einflussreiche Beobachtungen nach Cook's Distance}
\usage{
infl_obs(mod, threshold = 0.5, ...)
}
\arguments{
\item{mod}{lm oder gls Modell}

\item{threshold}{threshold at which influential observations are observed}

\item{...}{variable number of further arguments}
}
\value{
Liste mit dem Index der Beobachtungen und den Beobachtungen selbst
}
\description{
Alle Beobachtungen, deren Cook's distance größer als der Schwellenwert ist,
werden angegeben.
}
