% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infl_observations.R
\name{infl_observations.gls}
\alias{infl_observations.gls}
\title{Used in method infl_observations as gls model}
\usage{
\method{infl_observations}{gls}(
  mod,
  used_formula,
  cor_structure,
  threshold = 0.5,
  data,
  varname
)
}
\arguments{
\item{mod}{method used is gls}

\item{used_formula}{placeholder}

\item{cor_structure}{placeholder}

\item{threshold}{placeholder}

\item{data}{placeholder}

\item{varname}{placeholder}
}
\description{
Used in method infl_observations as gls model
}
