% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infl_observations.R
\name{infl_observations.lm}
\alias{infl_observations.lm}
\title{Used in method infl_observations as lm model}
\usage{
\method{infl_observations}{lm}(mod, threshold = 0.5, data, ...)
}
\arguments{
\item{mod}{method used is lm}

\item{threshold}{level of significance}

\item{data}{data which are used to fit the model}

\item{...}{placeholder}
}
\description{
Used in method infl_observations as lm model
}
