% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infl_observations.R
\name{infl_observations}
\alias{infl_observations}
\title{Define influential observations according to Cook's distance}
\usage{
infl_observations(mod, threshold = 0.5, ...)
}
\arguments{
\item{mod}{method lm or gls model}

\item{threshold}{Threshold at which observations are considered influential}

\item{...}{placeholder}
}
\value{
List with indices and observations
}
\description{
All observations whose Cook's distance is greater than the threshold
are specified#'
}
