% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymodelparm_default_imp.R
\name{mymodelparm_default_imp}
\alias{mymodelparm_default_imp}
\title{Unexported method for retrieving parameters from package preditmeans}
\usage{
mymodelparm_default_imp(model, coef. = coef, vcov. = vcov, df = NULL, ...)
}
\description{
Used in package trenda. Required in function CookD_gls
}
