% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot the trend model}
\usage{
plot_trend(mod, df, log_trans = FALSE)
}
\arguments{
\item{mod}{lm or gls model}

\item{df}{data}

\item{log_trans}{TRUE / FALSE for log-transformed data}
}
\value{
plot - object
}
\description{
Plots data as points and trends as lines
}
