% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bp.R
\name{ploti}
\alias{ploti}
\title{Plots the function trenda_bp() and compute_bp()}
\usage{
ploti(dat, varname, cuts, n.psi.ok, model.ok, dat1)
}
\arguments{
\item{dat}{data frame with observations}

\item{varname}{variable name within the data frame dat}

\item{cuts}{breakpoints determined by segmented in trenda_bp()}

\item{n.psi.ok}{numeric value indicationg the best breakpoints according to BIC}

\item{model.ok}{segmented model}

\item{dat1}{x and y values including the fitted segmented lines for plotting}
}
\value{
interactive plot using plotly and ggplot2
}
\description{
Uses ggplot2 and plotly to plot the results from trenda_bp()
}
