% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenda_dir.R
\name{trenda_dir}
\alias{trenda_dir}
\title{Perform trend analyses in time series related climate related data.}
\usage{
trenda_dir(
  data_dir,
  log_trans = FALSE,
  create_dir = TRUE,
  calc_infl_obs = TRUE
)
}
\arguments{
\item{data_dir}{Directory where data is stored. The function requires data to
be stored in a comma delimited csv file format and therefore only creates a
list of csv files from the specified folder.}

\item{log_trans}{logical. Set to TRUE if data is already log10-transformed.
If create_dir = TRUE a new folder named “_results_standard_log” will be created.
Observables will be transformed back to their original form and plots will be
adjusted for normal values.}

\item{create_dir}{logical. Creates a subfolder in the data_dir specified path
if set to TRUE. Print to console if set to FALSE If log_trans and calc_infl_obs
is set to TRUE the following folders are created:
“_results_standard”,
“_results_standard_obs”,
“_resulst_log”,
“_results_log_obs”}

\item{calc_infl_obs}{logical. TRUE removes the influential observations
generated and stored in the file “res_tab_standard_”, Also creates a new
folder named “_results_standard_obs”}
}
\value{
the function does not return a value but stores the plot and table in
the assigned directories
}
\description{
trenda_dir() is a use-case specific addition to the function trenda().
It was created for bulk-processing time series related data stored as
comma delimited csv files in a folder.
Each file is checked for
It performs a series of analysis to find trends
in time-series related data by first checking the number of observations,
performing the Generalized Durbin-Watson-Test to test for autocorrelation errors,
performing AR(1) or ARMA to get the correlation strucutre before peforming
a GLS
}
\details{
Provided data_dir all semicolon delimited csv files are analysed
and summarized in a csv file created in a subfolder. The first column of each
file is expected to be a time series of type int followed by any number of
columns with numeric values.

Based on the number of observations (n<7, 7 >= n < 13, n >= 13)
a linear equation or linear & quadratic equation is is formed.
A Durbin-Watson-Test with max.lag set to 2 is performed to find significant autocorrelation.
If p-value of first-order is smaller than  0.05 we reject H0 and assume significant autocorrelation
of first-order and perform an autoregressive model AR(1) to get the correlation structure for the GLS model.
If p >= 0.05 the second p-value of the DWT is inspected and checked again for significance.
If p >= 0.05 a linear regression model is fitted. For p < 0.05 we assume
autocorrelated  errors of second-order and perform an autoregressive-moving-average model
to get the correlation structure for GLS.
}
\examples{
# Assuming the folder with the data is placed in the working directory.
\dontrun{
list.files(path = "./files_folder")
#' # The analysis can be started by entering the path of the folder:
#' trenda_dir("./files_folder/", log_trans = FALSE, create_dir = TRUE, calc_infl_obs = TRUE)
}
}
