% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendanalyse_ohne_einfl_beob_function_2018.R
\name{trenda_obs}
\alias{trenda_obs}
\title{Function to calculate trend analyses for Bosch & Partner (without influential observations)}
\usage{
trenda_obs(
  data_dir,
  plot_dir,
  result_name,
  log_trans = FALSE,
  res_tab_file,
  calc_einfl_beob = TRUE
)
}
\arguments{
\item{data_dir}{directory where the data is stored}

\item{plot_dir}{directory where the generated plots are saved}

\item{result_name}{directory and name (without ending) where the result table
is stored}

\item{log_trans}{logical if log-transformed values are used (TRUE leads to
the adjustments for plot function)}

\item{res_tab_file}{result table of the same data of perform_trendanalyse;
needed for information about the influential observations which should be
left out in the analysis}

\item{calc_einfl_beob}{logical; the influential observations should be
calculated again. Can be useful to turn off as this information is not
needed and sometimes the calculation of the influential observations for
GLS models does not work}
}
\value{
the function does not return a value but stores the plot and table in
the assigned directories
}
\description{
Function to calculate trend analyses for Bosch & Partner (without influential observations)
}
\details{
Since this function needs the information which are influential obser-
vations, the function perform_trendanalyse has to be run first on the same
data.
this function needs the information generated by
}
