% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_trend.R
\name{compute_trend}
\alias{compute_trend}
\title{Calculates the trend for data and variables}
\usage{
compute_trend(dat, varname, log_trans = FALSE, calc_infl_obs = TRUE)
}
\arguments{
\item{dat}{Dataset}

\item{varname}{Columnames of dataset}

\item{log_trans}{logical gibt an, ob die zu modellierenden Daten log-transformiert sind}

\item{calc_infl_obs}{logical}
}
\value{
A list with trendgraphs, variablenames, model and information about the results
}
\description{
Calculates the trend for data and variables
}
