% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_trend.R
\name{compute_trend}
\alias{compute_trend}
\title{Calculates the trend for data and variables}
\usage{
compute_trend(dat, varname, log_trans = FALSE, calc_infl_obs = TRUE)
}
\arguments{
\item{dat}{Dataset}

\item{varname}{list with variablenames from the dataset}

\item{log_trans}{logical indicates if data is log-transformed}

\item{calc_infl_obs}{logical includes influential observations}
}
\value{
A list with trendgraphs, variablenames, model and information about the results
}
\description{
Checks whether enough data is available and runs the model
}
