% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phi.R
\name{extract_phi}
\alias{extract_phi}
\title{Extrahiere phi aus gls - Modell}
\usage{
extract_phi(mod)
}
\arguments{
\item{mod}{gls-Modell}
}
\value{
phi Parameter der Korrelation
}
\description{
Extrahiere phi aus gls - Modell
}
