% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phi.R
\name{extract_phi}
\alias{extract_phi}
\title{Extract phi}
\usage{
extract_phi(mod)
}
\arguments{
\item{mod}{gls model}
}
\value{
phi parameter of correlation
}
\description{
Helper function for the package trenda. Extracts phi value
}
