% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{Determine trend information using Sen's methodology and Kendall tests}
\usage{
get_trend(df, dt_col, val_col, start_window = NULL, window_width = NULL)
}
\arguments{
\item{df}{a data frame containing at least one column of date-time values, and a column of numeric data for which a trend analysis will be undertaken.}

\item{dt_col}{the name or column number of date-time values to be used in the analysis.}

\item{val_col}{the name or column number of numerical values to be used in the analysis.}

\item{start_window}{an integer representing the start of the windowed data used in the analysis.}

\item{window_width}{an integer representing the width of the windowed data used in the analysis.}
}
\value{
a data frame.
}
\description{
Using a vector of data, The Thiel-Sen method for trend determination will be carried out with a Kendall test for significance.
}

