% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymodelparm_default_imp.R
\name{mymodelparm_default_imp}
\alias{mymodelparm_default_imp}
\title{Unexported method for retrieving parameters from package preditmeans}
\usage{
mymodelparm_default_imp(model, coef. = coef, vcov. = vcov, df = NULL, ...)
}
\arguments{
\item{model}{model used in trenda}

\item{coef.}{coef used in trenda}

\item{vcov.}{calculated variance-covariance matrix for the fitted model object}

\item{df}{empty data frame}

\item{...}{further parameters}
}
\description{
Used in package trenda. Required in function CookD_gls
}
