% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenda.R
\name{trenda}
\alias{trenda}
\title{Function to perform trend analyses in time series and identify outliers using
Cook's Distance}
\usage{
trenda(df, plot_graphs = FALSE, log_trans = FALSE, calc_infl_obs = FALSE)
}
\arguments{
\item{df}{file where data is stored. The function requires data to
be stored in a semicolon delimited csv file format and therefore only creates a
list of csv files from the specified folder.}

\item{plot_graphs}{logical. Set to TRUE by default. Will plot to the environment.
The plots summarizes the information about the data the trend}

\item{log_trans}{logical. Set to TRUE if data is already log10-transformed.
If create_dir = TRUE a new folder named “_results_standard_log” will be created.
Observables will be transformed back to their original form and plots will be
adjusted for normal values.}

\item{calc_infl_obs}{logical: If set to TRUE the function will calculate the
influential observations using Cook's Distance. The data is stored in a list
with information}
}
\value{
If calc_infl_obs is set to true a list is returned containing two data.frames.
 The first data frame informs about the findings of the first run and suggested
 which values are influential using Cook's Distance. In a second run the influential
 observations are removed from the data frame and only those variables are reevaluated
 and returned in the second data frame.
}
\description{
performs a series of analysis to find trends
in time-series related data by first checking the number of observations,
performing the Generalized Durbin-Watson-Test to test for autocorrelation,
performing AR(1) or ARMA to get the correlation structure before performing
a GLS and further deciding on the trends of an observation by looking at significance
and beta values
}
\details{
The first column of the data frame is
expected to be a time series of type int followed by any number of
columns with numeric values.
Based on the number of observations (n<7, 7 >= n < 13, n >= 13)
a linear equation or linear & quadratic equation is is formed.
A Durbin-Watson-Test with max.lag set to 2 is performed to find significant autocorrelation.
If p-value of first-order is smaller than  0.05 we reject H0 and assume significant autocorrelation
of first-order and perform an autoregressive model AR(1) to get the correlation structure for the GLS model.
If p >= 0.05 the second p-value of the DWT is inspected and checked again for significance.
If p >= 0.05 a linear regression model is fitted. For p < 0.05 we assume
autocorrelated  errors of second-order and perform an autoregressive-moving-average model
to get the correlation structure for GLS.
}
\examples{
## Generate a data frame with to variables. One value of height_m is set
## to be extremely out of bound to illustrate how trenda() handles outliers.
\dontrun{ rnd_preci <- data.frame(Year = c(1991:2020), precip_mm = rnorm(30, 770, 50),
                               height_m = c(rnorm(10, 10, 1), 100,
                               rnorm(9, 15, 1),
                               rnorm(10, 20, 1)))

trenda(rnd_preci, plot_graphs = TRUE, log_trans = FALSE, calc_infl_obs = TRUE)
}
}
