% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenda.R
\name{trenda}
\alias{trenda}
\title{Function to perform trend analyses in time series}
\usage{
trenda(data_dir, log_trans = FALSE)
}
\arguments{
\item{data_dir}{directory where the data is stored}

\item{log_trans}{logical if log-transformed values are used (TRUE leads to
the adjustments for plot function)}
}
\value{
the function does not return a value but stores the plot and table in
the assigned directories
}
\description{
Given a folder with csv files
}
