% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenda_obs.R
\name{trenda_obs}
\alias{trenda_obs}
\title{Perform trend analyses on continous time series climate related data.}
\usage{
trenda_obs(
  data_dir,
  log_trans = FALSE,
  create_dir = TRUE,
  res_tab_file,
  calc_infl_obs = TRUE
)
}
\arguments{
\item{data_dir}{Directory where data is stored. The function requires data to
be stored in a comma delimited csv file format and therefore only creates a
list of csv files from the specified folder.}

\item{log_trans}{logical. Set to TRUE if data is already log10-transformed.
If create_dir = TRUE a new folder named “_results_standard_log” will be created.
Observables will be transformed back to their original form and plots will be
adjusted for normal values.}

\item{create_dir}{logical. Creates a subfolder in the data_dir specified path if set to TRUE. Print to console if set to FALSE
If log_trans and calc_infl_obs is set to TRUE the following folders are created:
 “_results_standard”
“_results_standard_obs”
“_resulst_log”
“_results_log_obs”}

\item{res_tab_file}{result table of the same data generated by trenda;
needed for information about the influential observations which should are
left out in this analysis}

\item{calc_infl_obs}{logical; the influential observations should be
calculated again. Can be useful to turn off as this information is not
needed and sometimes the calculation of the influential observations for
GLS models does not work}
}
\value{
the function does not return a value but stores the plot and table in
the assigned directories
}
\description{
Used to reevaluate the findings produced by trenda_dir() when
calc_infl_obs is set to FALSE. Should only be executed
after the function trenda_dir() was performed on data and influential observations
were determined. Also requires the ResTab csv file to work.
trenda_obs removes those influential observations and runs
the analysis again and places those findings
}
\details{
Since this function needs information about influential
observations, the function trenda has to be run first on the same
dataset.
}
