% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trenda_obs.R
\name{trenda_obs}
\alias{trenda_obs}
\title{Function to calculate trend analyses (without influential observations)}
\usage{
trenda_obs(
  data_dir,
  log_trans = FALSE,
  create_dir = TRUE,
  res_tab_file,
  calc_infl_obs = TRUE
)
}
\arguments{
\item{data_dir}{directory where the data is stored}

\item{log_trans}{logical if log-transformed values were used (TRUE leads to
the adjustments for plot function)}

\item{create_dir}{logical. Creates a subfolder in the data_dir specified path if set to TRUE. Print to console if set to FALSE
If log_trans and calc_infl_obs is set to TRUE the following folders are created:
 “_results_standard”
“_results_standard_obs”
“_resulst_log”
“_results_log_obs”}

\item{res_tab_file}{result table of the same data generated by trenda;
needed for information about the influential observations which should are
left out in this analysis}

\item{calc_infl_obs}{logical; the influential observations should be
calculated again. Can be useful to turn off as this information is not
needed and sometimes the calculation of the influential observations for
GLS models does not work}
}
\value{
the function does not return a value but stores the plot and table in
the assigned directories
}
\description{
Second main function of package trenda. Should only be executed
after the function trenda was performed on data and influential observations
were determined. trenda_obs removes those influential observations and runs
the analysis again
}
\details{
Since this function needs information about influential
observations, the function trenda has to be run first on the same
dataset.
}
